Support login with pin at runtime
When no AuthPass is provided and the AuthCert is password protected
a C_Login is required for a successful operation. In this
case the "Certificate" must be provided, because it can not be
retrieved initially without credentials.

The AuthCert may now also be a "PEM Cert + encrypted key"
instead of PKCS#12 (*.pfx)

During initialisation it tries to load the AuthCert and sets
the Session type (CKS_RO_USER_FUNCTIONS / CKS_PUBLIC_SESSION)
accordingly.